+++ EDDI/ChangeLog.md
@@ -1,5 +1,11 @@
 # CHANGE LOG
 
+### 2.4.0-b3
+  * Core
+    * Module events now update the ship object
+  * Events
+    * Revised variable names for module events
+
 ### 2.4.0-b2
   * Installer
     * Fixed: the installer was missing some of the documentation files. This was causing the app to pine for the fjords. And the documentation files.